+++ rest-hapi.js
@@ -11,6 +11,7 @@ const _ = require('lodash'),
     Q = require("q"),
     restHelperFactory = require('./utilities/rest-helper-factory'),
     handlerHelper = require('./utilities/handler-helper'),
+    joiHelper = require('./utilities/joi-mongoose-helper'),
     modelGenerator = require('./utilities/model-generator'),
     apiGenerator = require('./utilities/api-generator'),
     defaultConfig = require('./config');
@@ -35,7 +36,8 @@ module.exports = {
     getAll: handlerHelper.getAll,
     logger: {},
     getLogger: getLogger,
-    logUtil: logUtil
+    logUtil: logUtil,
+    joiHelper: joiHelper
 };
 
 function register(server, options, next) {
@@ -70,7 +72,8 @@ function register(server, options, next) {
                     title: config.appTitle,
                     version: config.version
                 },
-                expanded: config.docExpansion
+                expanded: config.docExpansion,
+                reuseDefinitions: false
             };
 
             server.register([