+++ tests/test-helper.js
@@ -23,7 +23,7 @@ module.exports = {
     parameters[logIndex] = Log;
 
     test(methodName + ' fails if model parameter isn\'t a mongoose model', function (t) {
-      t.plan(8);
+      t.plan(10);
 
       var model = {};
 
@@ -64,6 +64,23 @@ module.exports = {
         t.fail("No error was thrown.");
       } catch (error) {
         t.ok(/^AssertionError/.test(error.name), "error is an AssertionError");
+        t.ok(error.message.indexOf("schema.tree") > -1, "assertion message contains 'schema.tree' text.");
+      }
+
+      model = {
+        schema: {
+          paths: {},
+          tree: {}
+        }
+      };
+
+      paramCopy = _.extend([], parameters);
+      paramCopy[modelIndex] = model;
+      try {
+        methodToTest.apply(null, paramCopy);
+        t.fail("No error was thrown.");
+      } catch (error) {
+        t.ok(/^AssertionError/.test(error.name), "error is an AssertionError");
         t.ok(error.message.indexOf("routeOptions") > -1, "assertion message contains 'routeOptions' text.");
       }
 
@@ -75,7 +92,8 @@ module.exports = {
             field2: {},
             field3: {},
             field4: {}
-          }
+          },
+          tree: {}
         },
         routeOptions: {}
       };