+++ CHANGELOG.md
@@ -6,6 +6,24 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+## [0.26.0] - 2017-09-24
+### Added
+- Added config option to disable 500 response for response validation (disabled by default).
+
+### Changed
+- Cleaned up tests
+- Major updates to joi helper functions including support for nested fields.
+
+## [0.25.0] - 2017-09-20
+### Added
+- Added config option to enable logging of routes and scopes (disabled by default).
+
+## [0.24.0] - 2017-09-19
+### Added
+- Exposed [joi helper methods](https://github.com/JKHeadley/rest-hapi#joi-helper-methods), including the rest-hapi version of Joi ObjectId.
+### Fixed
+- Throw error when payload is empty for addMany.
+
 ## [0.23.0] - 2017-09-12
 ### Added
 - Allow custom errors in middleware
@@ -68,12 +86,15 @@ post(hardDelete, deleted, Log) | post(hardDelete, deleted, request, Log)
 
 
 
-[Unreleased]: https://github.com/jkheadley/rest-hapi/compare/v0.23.0...HEAD
+[Unreleased]: https://github.com/jkheadley/rest-hapi/compare/v0.26.0...HEAD
 [0.20.0]: https://github.com/jkheadley/rest-hapi/compare/v0.19.2...v0.20.0
 [0.20.2]: https://github.com/jkheadley/rest-hapi/compare/v0.20.0...v0.20.2
 [0.20.3]: https://github.com/jkheadley/rest-hapi/compare/v0.20.2...v0.20.3
 [0.21.0]: https://github.com/jkheadley/rest-hapi/compare/v0.20.3...v0.21.0
 [0.22.0]: https://github.com/jkheadley/rest-hapi/compare/v0.21.0...v0.22.0
 [0.23.0]: https://github.com/jkheadley/rest-hapi/compare/v0.22.0...v0.23.0
+[0.24.0]: https://github.com/jkheadley/rest-hapi/compare/v0.23.0...v0.24.0
+[0.25.0]: https://github.com/jkheadley/rest-hapi/compare/v0.24.0...v0.25.0
+[0.26.0]: https://github.com/jkheadley/rest-hapi/compare/v0.25.0...v0.26.0
 
 [BREAKING]: https://github.com/JKHeadley/rest-hapi/releases